Move SegmentFetcher to pinot-common from pinot-server
