+++ pinot-common/src/main/java/com/linkedin/pinot/common/segment/fetcher/SegmentFetcherFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.linkedin.pinot.server.segment.fetcher;
+package com.linkedin.pinot.common.segment.fetcher;
 
 import java.util.Arrays;
 import java.util.Iterator;
